Use a type handled by FR_TYPE_FROM_PTR

+++ src/lib/unlang/map_builtin.c
@@ -68,7 +68,7 @@ static unlang_action_t mod_list_map_proc(unlang_result_t *p_result, UNUSED map_c
 	rlm_rcode_t		rcode = RLM_MODULE_NOOP;
 	fr_value_box_t		*vb = NULL;
 	fr_value_box_t		**values;
-	size_t			index, i = 0, value_count = fr_value_box_list_num_elements(in);
+	uint32_t		index, i = 0, value_count = fr_value_box_list_num_elements(in);
 	TALLOC_CTX		*local = talloc_new(NULL);
 	map_t			*map = NULL;
 
@@ -92,11 +92,11 @@ static unlang_action_t mod_list_map_proc(unlang_result_t *p_result, UNUSED map_c
 			goto finish;
 		}
 		if (index > value_count) {
-			RWARN("Asked for index %ld when max is %ld.", index, value_count);
+			RWARN("Asked for index %d when max is %d.", index, value_count);
 			continue;
 		}
 		if (values[index]->type == FR_TYPE_NULL) {
-			RDEBUG2("Skipping null value for index %ld.", index);
+			RDEBUG2("Skipping null value for index %d.", index);
 			continue;
 		}
